界面填充

+++ app/src/main/java/com/chs/doubankotlin/module/bean/MovieDetail.kt
@@ -259,7 +259,7 @@ class MovieDetail{
          */
 
         var max: Int = 0
-        var average: Double = 0.toDouble()
+        var average: Float = 0.toFloat()
         var stars: String? = null
         var min: Int = 0
     }
